third_party/blink/web_tests/platform/generic/fast/forms/range/range-drag-expected.txt

Test for dragging operations of <input type=range>

Tests for range dragging from center.
readOnly=false, disabled=false
PASS input.value is "100"
PASS changeEventCounter is >= lastChangeEventCounter + 1
readOnly=true
PASS input.value is "100"
PASS changeEventCounter is >= lastChangeEventCounter + 1
disabled=true
PASS input.value is "50"
PASS lastChangeEventCounter is changeEventCounter


Tests for range dragging from edge
readOnly=false, disabled=false
PASS input.value is "100"
PASS changeEventCounter is >= lastChangeEventCounter + 1
readOnly=true
PASS input.value is "100"
PASS changeEventCounter is >= lastChangeEventCounter + 1
disabled=true
PASS input.value is "50"
PASS lastChangeEventCounter is changeEventCounter


PASS successfullyParsed is true

TEST COMPLETE
